REF: WalletView to handle all wallet views
